[12631] Use uint32 for phaseMasks in core and database

+++ src/game/Corpse.cpp
@@ -119,7 +119,7 @@ void Corpse::SaveToDB()
         << uint64(m_time) << ", "
         << uint32(GetType()) << ", "
         << int(GetInstanceId()) << ", "
-        << uint16(GetPhaseMask()) << ")";           // prevent out of range error
+        << uint32(GetPhaseMask()) << ")";           // prevent out of range error
     CharacterDatabase.Execute(ss.str().c_str());
     CharacterDatabase.CommitTransaction();
 }
