+++ src/game/GameObject.h
@@ -574,7 +574,7 @@ struct GameObjectData
 {
     uint32 id;                                              // entry in gamobject_template
     uint16 mapid;
-    uint16 phaseMask;
+    uint32 phaseMask;
     float posX;
     float posY;
     float posZ;
